Integrate cookie consent with Mixpanel initialization (off by default)
This adds a new option to Mixpanel initialization to integrate with Google
Tag Manager cookie consent.

This requires the user have already integrated Google Tag Manager with
a Consent Mode Provider:

https://support.google.com/tagmanager/answer/10718549?hl=en#cmp-integrations

If enabled, this will read if the user has granted `analytics_storage`
permission at initialization time. If not, it will opt out of Mixpanel.

In either case, it will listen to updates to cookie consent, and opt
the user out or in of Mixpanel as cookie consent is rescinded/granted
at runtime.